# Workaround to pass struct to kernel by ref for performance